Fix missing Q_OBJECT macros
Thanks to Clazy
