+++ src/core/qgsnetworkdiskcache.cpp
@@ -18,7 +18,9 @@
 
 #include "qgsnetworkdiskcache.h"
 
-QgsNetworkDiskCache::ExpirableNetworkDiskCache QgsNetworkDiskCache::sDiskCache;
+///@cond PRIVATE
+ExpirableNetworkDiskCache QgsNetworkDiskCache::sDiskCache;
+///@endcond
 QMutex QgsNetworkDiskCache::sDiskCacheMutex;
 
 QgsNetworkDiskCache::QgsNetworkDiskCache( QObject *parent )
