+++ src/core/raster/qgsrasterlayerrenderer.cpp
@@ -25,12 +25,50 @@
 #include "qgsproject.h"
 #include "qgsexception.h"
 
+
+///@cond PRIVATE
+
+QgsRasterLayerRendererFeedback::QgsRasterLayerRendererFeedback( QgsRasterLayerRenderer *r )
+  : QgsRasterBlockFeedback()
+  , mR( r )
+  , mMinimalPreviewInterval( 250 )
+{
+  setRenderPartialOutput( r->mContext.testFlag( QgsRenderContext::RenderPartialOutput ) );
+}
+
+void QgsRasterLayerRendererFeedback::onNewData()
+{
+  if ( !renderPartialOutput() )
+    return;  // we were not asked for partial renders and we may not have a temporary image for overwriting...
+
+  // update only once upon a time
+  // (preview itself takes some time)
+  if ( mLastPreview.isValid() && mLastPreview.msecsTo( QTime::currentTime() ) < mMinimalPreviewInterval )
+    return;
+
+  // TODO: update only the area that got new data
+
+  QgsDebugMsg( QString( "new raster preview! %1" ).arg( mLastPreview.msecsTo( QTime::currentTime() ) ) );
+  QTime t;
+  t.start();
+  QgsRasterBlockFeedback feedback;
+  feedback.setPreviewOnly( true );
+  feedback.setRenderPartialOutput( true );
+  QgsRasterIterator iterator( mR->mPipe->last() );
+  QgsRasterDrawer drawer( &iterator );
+  drawer.draw( mR->mPainter, mR->mRasterViewPort, mR->mMapToPixel, &feedback );
+  QgsDebugMsg( QString( "total raster preview time: %1 ms" ).arg( t.elapsed() ) );
+  mLastPreview = QTime::currentTime();
+}
+
+///@endcond
+///
 QgsRasterLayerRenderer::QgsRasterLayerRenderer( QgsRasterLayer *layer, QgsRenderContext &rendererContext )
   : QgsMapLayerRenderer( layer->id() )
   , mRasterViewPort( nullptr )
   , mPipe( nullptr )
   , mContext( rendererContext )
-  , mFeedback( new Feedback( this ) )
+  , mFeedback( new QgsRasterLayerRendererFeedback( this ) )
 {
   mPainter = rendererContext.painter();
   const QgsMapToPixel &qgsMapToPixel = rendererContext.mapToPixel();
@@ -231,34 +269,3 @@ QgsFeedback *QgsRasterLayerRenderer::feedback() const
   return mFeedback;
 }
 
-QgsRasterLayerRenderer::Feedback::Feedback( QgsRasterLayerRenderer *r )
-  : mR( r )
-  , mMinimalPreviewInterval( 250 )
-{
-  setRenderPartialOutput( r->mContext.testFlag( QgsRenderContext::RenderPartialOutput ) );
-}
-
-void QgsRasterLayerRenderer::Feedback::onNewData()
-{
-  if ( !renderPartialOutput() )
-    return;  // we were not asked for partial renders and we may not have a temporary image for overwriting...
-
-  // update only once upon a time
-  // (preview itself takes some time)
-  if ( mLastPreview.isValid() && mLastPreview.msecsTo( QTime::currentTime() ) < mMinimalPreviewInterval )
-    return;
-
-  // TODO: update only the area that got new data
-
-  QgsDebugMsg( QString( "new raster preview! %1" ).arg( mLastPreview.msecsTo( QTime::currentTime() ) ) );
-  QTime t;
-  t.start();
-  QgsRasterBlockFeedback feedback;
-  feedback.setPreviewOnly( true );
-  feedback.setRenderPartialOutput( true );
-  QgsRasterIterator iterator( mR->mPipe->last() );
-  QgsRasterDrawer drawer( &iterator );
-  drawer.draw( mR->mPainter, mR->mRasterViewPort, mR->mMapToPixel, &feedback );
-  QgsDebugMsg( QString( "total raster preview time: %1 ms" ).arg( t.elapsed() ) );
-  mLastPreview = QTime::currentTime();
-}
