+++ src/providers/arcgisrest/qgsamsdataitems.h
@@ -64,6 +64,8 @@ class QgsAmsConnectionItem : public QgsDataCollectionItem
 
 class QgsAmsLayerItem : public QgsLayerItem
 {
+    Q_OBJECT
+
   public:
     QgsAmsLayerItem( QgsDataItem *parent, const QString &name, const QString &url, const QString &id, const QString &title, const QString &authid, const QString &format );
 };
