+++ src/providers/grass/qgsgrassdatafile.h
@@ -34,6 +34,8 @@
  */
 class GRASS_LIB_EXPORT QgsGrassDataFile : public QFile
 {
+    Q_OBJECT
+
   public:
     explicit QgsGrassDataFile( QObject *parent = 0 );
     // We need FILE* to be able to test feof but QFile::open(FILE *, OpenMode) is not virtual
